Fix fetch data from API bug
