+++ src/styles/index.css
@@ -33,93 +33,54 @@ ul {
   align-items: center;
 }
 
+.comments-container li {
+  text-align: center;
+}
+
 #navbar li {
   padding: 12px;
   font-size: 18px;
   font-weight: 600;
 }
 
-.wrapper {
-  position: fixed;
-  width: 60%;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  -webkit-transform: translate(-50%, -50%);
-  -moz-transform: translate(-50%, -50%);
-  -ms-transform: translate(-50%, -50%);
-  -o-transform: translate(-50%, -50%);
-  background-color: rgb(66, 66, 108);
-  color: #fff;
-  border-radius: 16px;
-  -webkit-border-radius: 16px;
-  -moz-border-radius: 16px;
-  -ms-border-radius: 16px;
-  -o-border-radius: 16px;
-  display: none;
-  height: 90%;
-  overflow-y: scroll;
-  padding: 16px 0;
-}
-
-#close-btn {
-  position: absolute;
-  right: 32px;
-  top: 32px;
-  font-size: 24px;
-  background-color: #000;
-  width: 35px;
-  padding: 8px;
-  cursor: pointer;
-}
-
-.details img {
+.img-popup {
   width: 300px;
   margin: 5% 30%;
 }
 
-.details h1 {
-  text-align: center;
-  margin: 2% 0;
-}
-
-.details p {
-  text-align: center;
-  margin: 3%;
-  line-height: 28px;
-  font-size: 18px;
-}
-
-.comments h2 {
-  text-align: center;
-  margin-bottom: 14px;
-}
-
-.comments p {
+.subtitles-container p {
   text-align: center;
 }
 
-.add-comment h2 {
+.comments-container h3 {
   text-align: center;
-  margin: 12px 0;
+  margin-top: 10px;
+}
+
+.add-container h3 {
+  text-align: center;
 }
 
 form {
-  margin: 0 30%;
+  margin-left: 40%;
 }
 
 input[type="text"] {
   margin: 8px;
-  font-size: 18px;
-  font-weight: 600;
+  font-size: 16px;
+  font-weight: 400;
   padding: 10px;
 }
 
-.cards p {
-  text-align: start;
+.card {
   color: #fff;
-  height: auto;
-  margin: 10px;
+  border-radius: 10px;
+  background:
+    linear-gradient(
+      to right bottom,
+      rgba(255, 255, 255, 0.2),
+      rgba(22, 18, 21, 0.829)
+    );
 }
 
 #footer p {
@@ -134,44 +95,14 @@ input[type="text"] {
   border-radius: 10px;
 }
 
-.cards {
-  margin: 10px;
-  border: 2px solid #fff;
-  border-radius: 10px;
-  background:
-    linear-gradient(
-      to right bottom,
-      rgba(255, 255, 255, 0.2),
-      rgba(17, 2, 13, 0.829)
-    );
-}
-
 #image {
   width: 200px;
   margin: 10px 30px;
 }
 
-.heading {
-  width: 200px;
-  display: flex;
-  justify-content: space-between;
-  margin: 10px 30px;
-}
-
-.heading h2 {
+.btn {
+  border: 2px solid wheat;
   color: #fff;
-}
-
-.heading i {
-  color: red;
-  font-size: 24px;
-}
-
-button {
-  margin: 5px 50px;
-  font-size: 18px;
-  font-weight: 600;
-  padding: 8px;
 }
 
 .active {
